QQuickWidget: do not crash when failing to make context current
...in createFramebufferObject(), e.g. upon resizing.

Certain systems have a problem with OpenGL contexts during resizing,
according to user reports. We do not know why makeCurrent() would fail,
so we cannot prevent the failure, but we can check the result of
makeCurrent() and not go on doing GL stuff (which would almost certainly
crash) without a valid context.

Task-number: QTBUG-83319
Pick-to: 5.15
Change-Id: I50a4f4e06d636fa0ee41fb9b80e720500711854f
Reviewed-by: Andy Nichols <andy.nichols@qt.io>