Fix leftover argument names in ShaderEffect
"code" is misleading, and is a leftover from earlier times when the
value could have contained actual source code as well. We now have a
plain QUrl that is a local or qrc file reference. Update function
argument names accordingly. Also remove a confusing comment that was
again relevant when these properties were plain strings.

Change-Id: Ie88040d0c300a33167518098399dc0d1722a5961
Reviewed-by: Andy Nichols <andy.nichols@qt.io>