Give a better score for methods with a convertable type when matching
When it is looking for a matching method based on the argument types,
then if a QVariant can be converted to that type then it should give a
better score for that method. This is so that it can see it as being
more viable a choice when calling the method instead of potentially
not being able to find a matching one.

Change-Id: Ief7e11feacd1d0b0959330af2576c2d01affbc54
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit a056cb9595ea4a41c93f4c912719f9523b943d3b)
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>