CMake: Fix failing tst_qmlcachegen::skippedResources
The test was failing in CMake configurations because the pro2cmake
generated code set the QT_SKIP_QUICKCOMPILER property on a
non-existing file path (it was missing "data/").

Fix this manually by specifying the correct file path.

Add a SkipQmlCacheGen.qml file to make it more clear what
the file is used for. The previous alias from Retain.qml to Skip.qml
was confusing, and also was overriding the source file properties on
the same file that was part of 2 different resources.
Keep the alias to Skip.qml so we don't have to touch the C++ code.

pro2cmake will be fixed in a separate change to output correct file
paths.

Amends 41890402db086f4c61e052cef7659eb6b003dc77

Task-number: QTBUG-85364
Task-number: QTBUG-84886
Change-Id: I6e49deb8758403241d6f804af797adf2f05fda05
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>