Fix QQmlValueTypeWrapper::method_toString()
QMetaType::convert() is notoriously hard to use. You cannot keep the
QString on the stack as QString because its dtor will unconditionally be
called when the scope is left, no matter if the conversion succeeded.
Therefore, keep an array of quint8 on the stack and manually call the
dtor if the conversion succeeds.

Change-Id: If7970ff6adb4cc2a822a065f9f07130065ce9502
Reviewed-by: Lars Knoll <lars.knoll@qt.io>