Fix multiple windows with Vulkan on basic render loop
No problems with threaded, but when using QSG_RENDER_LOOP=basic with
Vulkan we missed setting the Vulkan instance for the second, third, etc.
window.

The problem can be seen with the quick/window example, running it with
QSG_RHI_BACKEND=vulkan QSG_RENDER_LOOP=basic makes it fail. With the
patch it will function as expected.

Change-Id: I98e7cb5ff960200dadb2fcbc30f771f9a7d9a9ae
Reviewed-by: Andy Nichols <andy.nichols@qt.io>