Fix tst_QQuickMenu::subMenuPosition()
We're still getting failures after QCOMPARE was fixed
(in c78a960198d59fb9a9ddd83ad7098b1db396edfd):

FAIL!  : tst_QQuickMenu::Material::subMenuPosition(cascading,flip) Compared doubles are not the same (fuzzy compare)
   Actual   (subSubMenu1->popupItem()->x())                              : 3.88240550819e-11
   Expected (subMenu1->popupItem()->x() - subSubMenu1->width() + overlap): 0
tst_qquickmenu.cpp(1532) : failure location

Quoting Eddy:

"[...] the threshold for double equality is one part in 1e12; the
threshold for fuzzy-is-zero is likewise 1e-12; so 3e-11 is non-zero.
One work-around, of course, would be to cast float(each operand) so as
to get float's wider tolerance (one part in 100,000, null if less than
1e-5)."

So that's what we do in this patch.

Change-Id: Iecf1b6f4b2cf2c81eb652bb0f565ac682b024dae
Pick-to: 5.15
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>