QQmlIRLoader: Restore inline components correctly
Whether a component is an inline component is not only stored in the
flags, but also in the isInlineComponent member.
Ideally, this should be unified and the member removed, but for now we
just restore the value correctly.
Adjusted tst_qmlcachegen::initTestCase so that we testFile and
testFileUrl are actually usable in the test.

Fixes: QTBUG-84237
Change-Id: I759cd6b8914b186b9e5c8118863fc8d0580d21af
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit b4d36a05be9f34976c1a1b5ea60ce03c4ab59a25)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>