Keep replaying touch press as mouse press after Flickable.pressDelay
In the future, if QQuickFlickable has pressDelay set, and receives a
touch event, we want it to replay the touch after the delay.  But in
Qt 5 it can only replay a delayed mouse press, and hacks are in place
to deal with that.  We aren't ready to change it all just yet, and
need to keep the autotest working as-is for now.

Task-number: QTBUG-78818
Change-Id: Ib37ccfc2b1b84254f40acac8f8ca8c51c9a88ddf
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>