Qt6QmlMacros.cmake: take the file path into consideration when copying QML files
When a QML plugin has QML files, and those files are
placed inside a directory (other than the root), qmldir
will also refer to those files prefixed with that directory.
Likewise, when the plugin is installed, the QML files will
need to copied so that they mirror this structure.
(This is also how qmake does it).

This patch will implement the same logic for cmake, so that
we check if the QML files are located elsewhere than the
root of the project, and if so, copy them into the same
location in the installation folder.

Change-Id: Iaebe4144d9b10b7d4922b53c546ca2f76bc0be1e
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>