Remove the pointId += deviceId << 24 hack from translateTouchEvent()
1204ed14dc11e6343a569646428a7ffddd098f8c did this in some places but
not others.  The result is mismatched pointIds in some cases: e.g.
a QObject event filter intercepts a touch press with modified pointIds
(because translateTouchEvent() got called) but the release event has the
original unmodified ids.

So we avoid this hack for touchpoint IDs, but keep it intact for mouse
and tablet points, for now.  The trouble is that 0 is a really common
value, so removing this hack everywhere can also break tests that are
checking simultaneous mouse and touch.

Eventually we need to start using the available QInputEvent::device() in
more places to avoid confusing points that come from different devices.

Fixes: QTBUG-85785
Change-Id: I873780b63e964c13dc96ed7fbf92cb019d079237
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>