Fix regression in antialiasing on outlined text
After e8d9bc1bcada7d94af5a33d64a9afc860ede5b84, outline text would
blend with the background rather than the fill color on the inner
side of the outline. We should do the mix with the fill color like
before this change, but for performance, we should still discard
all pixels which are not part of the outline, which we do using
step() instead of multiplying by 1.0 - a.

Pick-to: 5.15
Task-number: QTBUG-85514
Change-Id: I4b252bdbd9dae5ee599ba7c5d1cc3609fef61622
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>