Fix crash when QSGTexture is deleted in rc->endSync()
1. When the quick pixmap is cached, it might be deleted by a timer.
2. When its QQuickPixmapData is being deleted, its QQuickTextureFactory is also
deleted. Which informs the QSGRenderContext to postpone deleting of a texture
that belongs to this texture factory.
3. When an update is called on the _hidden_ window, QSGGuiThreadRenderLoop::renderWindow
calls rc->endSync(), which deletes postponed the texture from 2.

After that the texture must not be used.
But some QSGNode can still keep pointer to the deleted QSGTexture.
and when updatePaintNode is called, it might produce a crash.

So, suggesting a fix to inform the render loop that there is a window
with pending updates, and no need to delete textures.

Change-Id: I1487595dbb686e682ac3b91b9c3d21f401095daa
Fixes: QTBUG-65170
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 3df5ad9d247b194401d61c3ffb4b816ec239bded)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>