Ship with optimized SPIR-V for the built-in shaders
Also includes stripping the shaders properly. While we do not have debug
information, we do have OpName and similar that describe the variable and
block member names. This is unnecessary in the final SPIR-V binary, and further
reduces the size somewhat.

To create debuggable shaders (for SPIR-V and D3D) with full source information,
one needs to regenerate the relevant ones with -g instead of -O.

Change-Id: I3b855ee4e9afe393cd460407c1a807d6c9c5e6c2
Reviewed-by: Andy Nichols <andy.nichols@qt.io>