doc, listview: document that binding to parent should be avoided
Clearify in the documentation that the parent of a delegate
can change during its lifetime, and even be null. And because
of that, they should avoid binding to it.

In general, how a control is structured internally should be
understood as a private implementation detail. Even if the
application has the freedom to read (and even change) the
internal parent-child hierarchy as it sees fit, doing so
can be fragile as the internal structure of a control
can easily change from one version to the other.

[ChangeLog][QtQuick][ListView] Since the parent of a delegate can
end up being null during its lifetime, developers are
now advised against using it in bindings.

Fixes: QTBUG-82989
Change-Id: I2d6c79102435d0969f4dc3deec13c5fff4affeb1
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit ed73efa27acd8e7c42a960ef90c2e4a898c20c30)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>