CMake: Don't find_package(Qt6QmlTools) directly for cross-builds
It will fail when cross compiling to a platform that has a
different size value of CMAKE_SIZEOF_VOID_P than the host one.

Instead call find_package() on the base package (Qml in this case) and
it will look for QmlTools itself in a special way that that resets
the value of CMAKE_SIZEOF_VOID_P, thus ensuring the package is found.

Fixes: QTBUG-86012
Change-Id: Ibf7f06b8cfea0ccde10b7ca8429ef8ae598f93f2
Reviewed-by: Cristian Adam <cristian.adam@qt.io>