Fix wrong optimization on ARM64 ubuntu qemu
in QQuickDesignerSupportProperties::propertyNameListForWritableProperties
inspected object is initialized with a stack variable in the first
iteration:

    if (inspectedObjects == nullptr)
        inspectedObjects = &localObjectList;

Subsequently inspectedObjects->contains(object) is called.
This crashes on ubuntu ARM64 QEMU unless one performs other operations
before that call (just in optimized code without debug information).
As the contains will return false in the first iteration doing it only
in subsequent recursions (by putting it in an else clause) does not
change the expected execution, but avoid the optimization bug that was
blocking CI integration of qdeclarative (a crash cannot be blacklisted).

Change-Id: I2d14ecc78673ad06073825c094e4e1dcd935e73a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>