CMake: Adjust naming of Qt Qml plugins for Android
The names are identical to the qmake ones now, following the scheme
that the plugin file name should contain the TARGET_PATH, aka the
directory path of the plugin, encoded in the file name
itself because that's how the Android Qml plugin loader expects them.

A second change is needed to pass the correct library suffix as the
define, so that the Qml plugin loader actually finds them.

After that, the following error goes away.

libemitters_x86: qrc:/particles/emitters/emitters.qml:51:1:
         module "QtQuick" plugin "qtquick2plugin" not found
         import QtQuick 2.0

Task-number: QTBUG-85399
Change-Id: I16b9f877db3e5c8266429772a44231f1a356515c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>