QQuickLoader: make sure the status property change properly
When we call [setSource] in qml, we will try to convert a qml value to
a string which indicate the url we want to load. However, if this value
is undefined, it will convert to a string with content "undefined", and
then cause loading error. More worse, if there is a qml file named
"undefined", it will be loaded to the Loader.

Pick-to: 5.15
Fixes: QTBUG-85938
Change-Id: I5b192ba84aa29532e547349bbf37d413c107d8e4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>