CMake: Fix what Android lib suffix should be passed as a define
CMAKE_SYSTEM_PROCESSOR has the value i686 when targeting the x86
Android ABI, which is set by the Android CMake toolchain.

This is not the value that Qt expects, and breaks loading of Qt Qml
plugins. The same with aarch64 instead of arm64-v8a.

In qmake the proper values for QT_ARCH variable are the equivalent of
the ANDROID_ABI variable set by the CMake toolchain file, so use
ANDROID_ABI.

This allows building and running a qtdeclarative example built with
CMake on the x86 Android emulator.

Note that the x86 emulator still fails to actually run the examples
properly due to some weird atomics consistency issues, but that's a
separate issue. At least it doesn't fail saying that the Qt Quick 2
plugin can not be found.

Task-number: QTBUG-85399
Change-Id: Ic85c7f998b6d69793e4f7f0408fca6bdc88f6416
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>