Adapt to QNetworkRequest change
In d12afeff4097dd78f1311511e0c8f74e1a810d47 the default for
redirection handling was changed from manual to
NoLessSafeRedirectPolicy.
This makes the redirection handling asynchronous, leading to a failure
in tst_qqmlxmlhttprequest::redirects.
In it the request is not finished after component.completeCreate().
Thus the property dataOK is read before being set to true, and the
test fails.
This change reestablishes the ManualRedirectPolicy, giving the old
behavior, and fixes a failure blocking CI.

Change-Id: I525b1229b4ce8cfb814bdb66f86f745ea05c1b5c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>