tst_qqmlqt: Avoid warning in test
The test files still contained a property to test functionality which
was in the end discarded; thus passing the wrong number of arguments to
functions and causing a warning. We simply remove them as they are not
needed anywhere.
Amends 19850f129881a04c36c51859454bf71ed688f28a
As a drive by, replace the obsolete variant with var.

Change-Id: If0b3dcd6d8a339a569c39ce304b68fecd3d72779
Reviewed-by: Maximilian Goldstein <max.goldstein@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>