Fix crash when importing a style without first importing Controls
The following code

    import QtQuick
    import QtQuick.Controls.Material

    ApplicationWindow {
        width: 200
        height: 200
        visible: true
    }

produced an error in Qt 5:

    QQmlApplicationEngine failed to load component
    qrc:/main.qml:4 ApplicationWindow is not a type

In Qt 6, the types are provided by the qmldir, so the import will work,
but as QtQuickControls2Plugin has not been loaded, there is no
QQuickTheme object yet, and so the style will not work as expected
(any colors, fonts, etc. from the theme will not be used by the
style).

Produce a warning for this scenario, and test each style to make sure
that we don't crash.

Fixes: QTBUG-86280
Change-Id: I99f940255f56da0522ad192ae5da4c9110ea308e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>