Revert "QQuickWidget: Propagate focusObjectChanged"
This reverts commit 9bfa722bc0b67e01f35d78e1165602e174ae3b02.

During clean up of the patch, a bug was introduced which actually
causes it to have no effect (except print out a warning), because
the focusObjectChanged() signature is wrong. When the bug is fixed,
it turns out that it actually breaks the QQuickWidget autotest.

So this may have been a bit hasty. Reverting it and returning
to the drawing board.

Task-number: QTBUG-61475
Change-Id: Iac4507ee4c92d94b0db5b996dfa2206be564cfb3
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>