Propagate focusObjectChanged signal (second attempt)
When the focus object is updated from inside the Qt Quick scene,
the signal needs to be propagated from the offscreen window to
the widget's window, otherwise the input methods will not react
to it.

Also, we need to propagate the FocusAboutToChange event,
otherwise the contents of the editor will not be committed
when the focus object changes, and the input method can
get into an invalid state.

Fixes: QTBUG-61475
Change-Id: I44ba171c0e78ef8b2e0127cba8991f1f1cf13571
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 362f6999a52849482e73915ca37ed4b6cc9ff9aa)