Follow up on move of resolveMask into QPalettePrivate
The old QPalette move constructor and assignment operators only exchanged
the d-pointer, but left the resolveMask of the moved-to palette instance
untouched.

This was clearly wrong. With the move of the resolveMask to the private in
556511f9f39ddc887481e0cd5a877134ceb0da6b, the resolveMask is moved, as one
would expect.

However, this broke the inheritance of palettes in QtQuickControls2, as
QQuickPaletteColorProvider::inheritPalette did not restore the mask of the
resulting palette to the mask of the palette that inherited from.

Change-Id: I7436d3e493e06b8619102212ac7b707daa973b2a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Vitaly Fanaskov <vt4a2hqt@gmail.com>