Render loops: ignore update reqs during the main polish step
The classic example is Shape, which needs to dirty the QQuickItem in
updatePolish() in order to get it picked up in the synchronize step.
That part is fine, but we do not want maybeUpdate() to issue a
requestUpdate() then since we are effectively in progress of doing an
update, so having another full round of polish/sync/render is a waste.

Task-number: QTBUG-86089
Change-Id: Ie41563b34da17e7134631791ed024b31e87e21e3
Reviewed-by: Andy Nichols <andy.nichols@qt.io>