CMake: Make reconfiguration work for in-source builds
Qml plugins can either use a plugins.qmltypes file from the source
repository, or opt into generating it if the file doesn't exist.
There is a use case for the file not to be generated for some Qml
tests.

To protect against accidentally overriding the source repo file,
an error is shown if the file exists in the source directory and we
also also try to generate it.

With an in-source build, all plugins.qmltypes files are generated in
the source directory, and this causes the above check to be triggered
on reconfiguration.

To avoid the false issue, create a new
"${CMAKE_CURRENT_BINARY_DIR}/.generated/plugins.qmltypes" file to
serve as a marker that the actual file was generated, and is not part
of the source repository.

If the marker file exists, don't error out.

Change-Id: Ieca28b4da0251f67f8c066cbac471c8e6e955758
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>